calm.profile_api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
stripe==8.11.0
Jinja2==3.1.4
WeasyPrint==62.3
sendgrid==6.11.0
gunicorn==21.2.0